Restructured the flash API and implemented STM32 flash page erasing
Writes to flash are now a three-step process: begin(), fill()..., and end().
The API is structured this way so that differences between AVRs and
STM32s are minimized. On STM32 data is directly written to the target page; on
AVR it will be written to the internal scrap page and then copied over.