Modified CAN messages according to the new specs
The new system should make it easier to accept writes or rollback a
whole page of flash at a time.

TODO: The next page can probably be written to while flash is busy
COMMITting a previously-written page. WRITEs will just go to a temporary
location in RAM and not to the destination itself until a COMMIT
anyways...