Restructured the build system for the STM32 lib
Now you just have to specify a toolchain file from CANnuccia's source
code and CMake will figure out which target to build for.
For instance, if you are in a `build/stm32/` directory from the project's
root, you can run:
cmake -GNinja
      -DCMAKE_TOOLCHAIN_FILE=../../src/stm32/STM32toolchain.cmake
to make CANnuccia be built for STM32. The same will be done for AVR,
with a specific toolchain file.